+++ examples/waves/src/shaders/ocean.ts
@@ -83,6 +83,5 @@ vec3 calcColor() {
 
 void main() {
   csm_DiffuseColor = vec4(calcColor(), 1.0);   
-
 }
     `