+++ src/keywords.ts
@@ -1,5 +1,6 @@
 export default {
-  posiiton: 'csm_Position',
+  positon: 'csm_Position',
+  emissive: 'csm_Emissive',
   normal: 'csm_Normal',
   pointSize: 'csm_PointSize',
   diffuseColor: 'csm_DiffuseColor',