+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 3.2.10
+
+### Changes
+
+- Now supports `csm_Emissive`
+  - Override emissive color from fragment shader
+- Updated types to include `PointsMaterial` (#15)
+
 ## Version 3.1.0 **[Breaking]**
 
 ### Changes